Refactor freezing mechanics in entities
